pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "serialmidi-everdrive"
version = "0.1.0"
description = "Serial MIDI bridge with Mega Everdrive PRO support"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [{ name = "Ryan Kojima" }, { name = "Robert Hargreaves" }]
keywords = ["midi", "serial", "everdrive", "sega", "mega-drive", "genesis"]
classifiers = [
	"Development Status :: 4 - Beta",
	"Environment :: Console",
	"License :: OSI Approved :: Apache Software License",
	"Programming Language :: Python :: 3",
	"Topic :: Multimedia :: Sound/Audio :: MIDI",
]
dependencies = ["python-rtmidi", "pyserial"]

[project.urls]
Homepage = "https://github.com/rhargreaves/serialmidi"
Repository = "https://github.com/rhargreaves/serialmidi"

[project.scripts]
serialmidi = "serialmidi:main"
